pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "healthchecks-utilities"
version = "0.1.0"
description = "Utilities for monitoring domain health and expiry using Healthchecks.io"
readme = "README.md"
requires-python = ">=3.6"
license = {file = "LICENSE"}
authors = [
    {name = "Healthchecks Utilities Contributors"}
]
dependencies = [
    "requests==2.32.3",
    "python-dotenv==1.1.0",
]

[tool.setuptools]
packages = ["domain_checker"]
